aidecamp-a2a/webapi/pyproject.toml
[project]
name = "aidecamp-a2a"
version = "0.1.0"
description = "AI de Camp wargaming application with A2A protocol support using Semantic Kernel"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "semantic-kernel>=1.30.0",
    "a2a-sdk>=0.2.9",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "pillow>=10.0.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "starlette>=0.27.0",
    "click>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]